chore: move attributeClass service to new folder structure (#944)
* chore: move attributeClass service

* update person service with new attributeClass location
