+++ apps/web/app/(app)/environments/[environmentId]/(actionsAndAttributes)/attributes/AttributeSettingsTab.tsx
@@ -5,7 +5,7 @@ import type { AttributeClass } from "@prisma/client";
 import { useForm } from "react-hook-form";
 import { ArchiveBoxArrowDownIcon, ArchiveBoxXMarkIcon } from "@heroicons/react/24/solid";
 import { useRouter } from "next/navigation";
-import { updatetAttributeClass } from "@formbricks/lib/services/attributeClass";
+import { updatetAttributeClass } from "@formbricks/lib/attributeClass/service";
 import { useState } from "react";
 
 interface AttributeSettingsTabProps {
