+++ apps/web/app/(app)/environments/[environmentId]/(actionsAndAttributes)/attributes/page.tsx
@@ -5,7 +5,7 @@ import AttributeClassDataRow from "@/app/(app)/environments/[environmentId]/(act
 import AttributeTableHeading from "@/app/(app)/environments/[environmentId]/(actionsAndAttributes)/attributes/AttributeTableHeading";
 import HowToAddAttributesButton from "@/app/(app)/environments/[environmentId]/(actionsAndAttributes)/attributes/HowToAddAttributesButton";
 import { REVALIDATION_INTERVAL } from "@formbricks/lib/constants";
-import { getAttributeClasses } from "@formbricks/lib/services/attributeClass";
+import { getAttributeClasses } from "@formbricks/lib/attributeClass/service";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
