+++ apps/web/app/api/v1/management/attribute-classes/route.ts
@@ -4,7 +4,7 @@ import { authenticateRequest } from "@/app/api/v1/auth";
 import { NextResponse } from "next/server";
 import { transformErrorToDetails } from "@/lib/api/validator";
 import { TAttributeClass, ZAttributeClassInput } from "@formbricks/types/v1/attributeClasses";
-import { createAttributeClass, getAttributeClasses } from "@formbricks/lib/services/attributeClass";
+import { createAttributeClass, getAttributeClasses } from "@formbricks/lib/attributeClass/service";
 
 export async function GET(request: Request) {
   try {
